requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile pyproject.toml
#
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
idna==3.4
    # via requests
pyyaml==6.0
    # via os-mfa (pyproject.toml)
requests==2.29.0
    # via os-mfa (pyproject.toml)
urllib3==1.26.15
    # via requests
